fix: add message object to request and response shims and organise grpc under a directory
